Refactor: VXSEデバッグeditor controllerのトップレベル関数をValidator/IdentityGenerator/LevelMoverへ切り出し
